# openapi.yaml
openapi: 3.0.3
info:
  title: Monumental Link Manager API
  description: Production URL shortener with click tracking for monumental-i.com organization
  version: 1.0.0
  contact:
    name: Monumental I
    email: admin@monumental-i.com
servers:
  - url: https://us-central1-moni-url-short.cloudfunctions.net/api
    description: Production server
  - url: http://localhost:5001/moni-url-short/us-central1/api
    description: Local development server
security:
  - FirebaseAuth: []
paths:
...
  /api/v1/links:
    post:
      summary: Create a new short link
      description: Creates a new short link. Requires authentication with @monumental-i.com email.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - longUrl
                - createdBy
              properties:
                slug:
                  type: string
                  description: Custom slug (optional, auto-generated if not provided)
                  example: my-custom-slug
                  minLength: 1
                  maxLength: 50
                longUrl:
                  type: string
                  format: uri
                  description: The long URL to shorten
                  example: https://example.com/very-long-url
                createdBy:
                  type: string
                  format: email
                  description: Email of the user creating the link
                  example: user@monumental-i.com
                disabled:
                  type: boolean
                  description: Whether the link is disabled
                  default: false
                notes:
                  type: string
                  description: Optional notes about the link
                  example: Marketing campaign link
                tags:
                  type: array
                  items:
                    type: string
                  description: Optional tags for categorization
                  example: ["marketing", "campaign"]
                emailAlerts:
                  type: boolean
                  description: Enable Google Chat alerts for this link
                  default: false
      responses:
        '201':
          description: Link created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Link'
                  message:
                    type: string
                    example: Link created successfully
        '400':
          description: Invalid input data
        '401':
          description: Authentication required
        '403':
          description: Access restricted to monumental-i.com organization

    get:
      summary: List all links
      description: Retrieves a paginated list of links. Requires authentication with @monumental-i.com email.
      parameters:
        - name: limit
          in: query
          description: Number of links to return (max 100)
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          description: Number of links to skip
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: search
          in: query
          description: Search term for slug or URL
          schema:
            type: string
      responses:
        '200':
          description: List of links retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Link'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
  /api/v1/links/{slug}:
    get:
      summary: Get a specific link
      description: Retrieves details for a specific link by slug. Requires authentication with @monumental-i.com email.
      parameters:
        - name: slug
          in: path
          required: true
          description: The link slug
          schema:
            type: string
            example: my-custom-slug
      responses:
        '200':
          description: Link details retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Link'
        '404':
          description: Link not found

    patch:
      summary: Update a link
      description: Updates an existing link. Requires authentication with @monumental-i.com email.
      parameters:
        - name: slug
          in: path
          required: true
          description: The link slug
          schema:
            type: string
            example: my-custom-slug
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                longUrl:
                  type: string
                  format: uri
                  description: Updated long URL
                disabled:
                  type: boolean
                  description: Whether the link is disabled
                notes:
                  type: string
                  description: Updated notes
                tags:
                  type: array
                  items:
                    type: string
                  description: Updated tags
                emailAlerts:
                  type: boolean
                  description: Enable/disable Google Chat alerts
      responses:
        '200':
          description: Link updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Link'
                  message:
                    type: string
                    example: Link updated successfully

    delete:
      summary: Delete a link
      description: Deletes a link permanently. Requires authentication with @monumental-i.com email.
      parameters:
        - name: slug
          in: path
          required: true
          description: The link slug
          schema:
            type: string
            example: my-custom-slug
      responses:
        '200':
          description: Link deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Link deleted successfully
  /api/v1/links/{slug}/clicks:
    get:
      summary: Get click logs for a link
      description: Retrieves click logs for a specific link. Requires authentication with @monumental-i.com email.
      parameters:
        - name: slug
          in: path
          required: true
          description: The link slug
          schema:
            type: string
            example: my-custom-slug
        - name: from
          in: query
          description: Start date for click logs (ISO 8601)
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          description: End date for click logs (ISO 8601)
          schema:
            type: string
            format: date-time
        - name: limit
          in: query
          description: Number of clicks to return (max 200)
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
        - name: offset
          in: query
          description: Number of clicks to skip
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Click logs retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Click'

    delete:
      summary: Clear all click logs for a link
      description: Deletes all click logs for a specific link and resets click count. Requires authentication with @monumental-i.com email.
      parameters:
        - name: slug
          in: path
          required: true
          description: The link slug
          schema:
            type: string
            example: my-custom-slug
      responses:
        '200':
          description: Click logs cleared successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      deletedCount:
                        type: integer
                        example: 15
                  message:
                    type: string
                    example: Cleared 15 click logs
  /api/v1/settings:
    get:
      summary: Get global settings
      description: Retrieves global system settings. Requires authentication with @monumental-i.com email.
      responses:
        '200':
          description: Settings retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      globalEmailAlerts:
                        type: boolean
                        description: Global email alerts setting
                        example: false
        '401':
          description: Authentication required
        '403':
          description: Access restricted to monumental-i.com organization

    patch:
      summary: Update global settings
      description: Updates global system settings. Requires authentication with @monumental-i.com email.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                globalEmailAlerts:
                  type: boolean
                  description: Enable/disable global email alerts
                  example: true
      responses:
        '200':
          description: Settings updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      globalEmailAlerts:
                        type: boolean
                        example: true
                  message:
                    type: string
                    example: Settings updated successfully
...
components:
  securitySchemes:
    FirebaseAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Firebase Authentication token
  schemas:
    Link:
      type: object
      properties:
        id:
          type: string
          example: my-custom-slug
        slug:
          type: string
          example: my-custom-slug
        longUrl:
          type: string
          format: uri
          example: https://example.com/very-long-url
        createdAt:
          type: string
          format: date-time
          example: "2025-09-20T02:45:23.123Z"
        createdBy:
          type: string
          format: email
          example: user@monumental-i.com
        disabled:
          type: boolean
          example: false
        clickCount:
          type: integer
          example: 42
        lastClickedAt:
          type: string
          format: date-time
          nullable: true
          example: "2025-09-20T02:45:23.123Z"
        notes:
          type: string
          nullable: true
          example: Marketing campaign link
        tags:
          type: array
          items:
            type: string
          nullable: true
          example: ["marketing", "campaign"]
        emailAlerts:
          type: boolean
          example: true
    Click:
      type: object
      properties:
        id:
          type: string
          example: click-123
        slug:
          type: string
          example: my-custom-slug
        ts:
          type: string
          format: date-time
          example: "2025-09-20T02:45:23.123Z"
        ip:
          type: string
          example: 192.168.1.100
        userAgent:
          type: string
          example: Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36
        referer:
          type: string
          nullable: true
          example: https://example.com/page
        hostname:
          type: string
          nullable: true
          example: example.com
        country:
          type: string
          nullable: true
          example: United States
        region:
          type: string
          nullable: true
          example: Missouri
        city:
          type: string
          nullable: true
          example: Kansas City
        timezone:
          type: string
          nullable: true
          example: America/Chicago
        isp:
          type: string
          nullable: true
          example: Comcast Cable
    Pagination:
      type: object
      properties:
        total:
          type: integer
          example: 150
        limit:
          type: integer
          example: 20
        offset:
          type: integer
          example: 0
        hasMore:
          type: boolean
          example: true